Add doctrine migrations and move news to database
