+++ public/news/index.php
@@ -20,8 +20,8 @@
                 <arg name="format">
                     <article class="mb-5">
                         <h3>{{title}}</h3>
-                        <p>
-                            <i>Published: <date>{{date}}</date></i>
+                        <p class="text-muted">
+                            <date>{{publish_date}}</date>
                         </p>
                         {{body}}
                         <hr/>
