+++ src/Entity/News.php
@@ -36,21 +36,21 @@ class News
     /**
      * @var DateTime
      *
-     * @ORM\Column(name="publish_date", type="datetime", nullable=true})
+     * @ORM\Column(name="publish_date", type="datetime", nullable=true)
      */
     private $publishDate;
 
     /**
      * @var DateTime
      *
-     * @ORM\Column(name="display_start_date", type="datetime", nullable=true})
+     * @ORM\Column(name="display_start_date", type="datetime", nullable=true)
      */
     private $displayStartDate;
 
     /**
      * @var DateTime
      *
-     * @ORM\Column(name="display_end_date", type="datetime", nullable=true})
+     * @ORM\Column(name="display_end_date", type="datetime", nullable=true)
      */
     private $displayEndDate;
 
@@ -64,7 +64,7 @@ class News
     /**
      * @var string body may contain HTML
      *
-     * @ORM\Column(name="body", type="blob", length=65535, nullable=false)
+     * @ORM\Column(name="body", type="text", length=65535, nullable=false)
      */
     private $body;
